Simplify CreateNPCTrainerParty & rework SetWildMonHeldItem
Simplified CreateNPCTrainerParty so it could in theory just be used to populate enemy party data with the party data set to a NPC trainer.

Reworked SetWildMonHeldItem so it's able to set a wild mon item to a specified party slot of the enemy party during a wild battle. Useful for if wild doubles get implemented.